+++ CHANGES.md
@@ -55,6 +55,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## 1.0.0 - 2025-08-06
 ### New
-- Initial release of CheckMK MKP Builder Action
+- Initial release of Checkmk MKP Builder Action